backend/requirements/production.txt
flask==2.1.1
flask-login==0.6.0
flask-migrate==3.1.0
flask-sqlalchemy==2.5.1
flask-socketio==5.1.1
flask-wtf==1.0.1
flask-restful==0.3.9
email-validator==1.1.3
eventlet==0.33.1
psycopg2==2.9.3
https://github.com/benoitc/gunicorn/archive/refs/heads/master.zip#egg=gunicorn==20.1.0